vm/adb/adb.go: Updates to support boards that may lack a battery.
This became needed because the board on which ARM32 was tested did not
have a battery. That was causing a failure with original adb.go since
the code looked for the battery level to be sufficiently high. Now, the
code first checks if there is a battery service. If so, it reverts to
the old check. Else, it ignores the battery level.